auth

+++ todo app oop/delete.php
@@ -1,4 +1,9 @@
 <?php 
+ session_start();
+    if(!isset($_SESSION['username'])){
+        header('location:./login.php');
+        exit();
+    }
     include './classes/todo.php';
     $todo=new todo;
     $todo->delete($_GET['id']);
